Moved NFC to new class
